Move getFileBucketAndKey to common
