+++ packages/api/endpoints/dashboard.js
@@ -2,28 +2,14 @@
 
 const router = require('express-promise-router')();
 const {
-  aws
+  getS3Object,
+  getFileBucketAndKey
 } = require('@cumulus/common');
 
-/*
- * To Do: Took this from dist, move to common
- */
-function getFileBucketAndKey(pathParams) {
-  const fields = pathParams.split('/');
-
-  const Bucket = fields.shift();
-  const Key = fields.join('/');
-
-  if (Bucket.length === 0 || Key.length === 0) {
-    throw new Error(pathParams);
-  }
-
-  return [Bucket, Key];
-}
-
-
 /**
- * TO DO
+ * Given a path in the form of bucket/key, get the item from
+ * S3 and return it with the corret content type. This is used to server
+ * the dashboard from the specified bucket
  *
  * @param {Object} req - express request object
  * @param {Object} res - express response object
@@ -32,7 +18,7 @@ function getFileBucketAndKey(pathParams) {
 async function get(req, res) {
   const params = getFileBucketAndKey(req.params[0]);
 
-  const getObjectResponse = await aws.getS3Object(params[0], params[1]);
+  const getObjectResponse = await getS3Object(params[0], params[1]);
   res.set('Content-Type', getObjectResponse.ContentType);
   return res.send(getObjectResponse.Body.toString());
 }
